+++ test/postgresql_test.rb
@@ -7,4 +7,8 @@ class TestPostgresql < Minitest::Test
     super
     User.establish_connection adapter: "postgresql", database: "groupdate_test"
   end
+
+  def test_no_column
+    assert_raises(ArgumentError) { User.group_by_day.first }
+  end
 end